refactor: change adapter name share req -shortcode
