+++ packages/hoppscotch-common/src/helpers/backend/gql/mutations/CreateShortcode.graphql
@@ -2,5 +2,7 @@ mutation CreateShortcode($request: String!) {
   createShortcode(request: $request) {
     id
     request
+    createdOn
+    properties
   }
 }
