+++ packages/hoppscotch-common/src/helpers/backend/gql/queries/GetMyShortcodes.graphql
@@ -3,5 +3,6 @@ query GetUserShortcodes($cursor: ID) {
     id
     request
     createdOn
+    properties
   }
 }
