+++ packages/hoppscotch-common/src/helpers/backend/gql/subscriptions/ShortcodeCreated.graphql
@@ -3,5 +3,6 @@ subscription ShortcodeCreated {
     id
     request
     createdOn
+    properties
   }
 }
