--- packages/hoppscotch-common/src/helpers/sharedRequest/SharedRequest.ts
@@ -1,8 +1,0 @@
-/**
- * Defines how a Shared Request is represented in the SharedRequestListAdapter
- */
-export interface SharedRequest {
-  id: string
-  request: string
-  createdOn: Date
-}
