+++ packages/hoppscotch-common/src/helpers/shortcode/Shortcode.ts
@@ -1,0 +1,9 @@
+/**
+ * Defines how a Shortcode is represented in the ShortcodeListAdapter
+ */
+export interface Shortcode {
+  id: string
+  request: string
+  properties?: string | null | undefined
+  createdOn: Date
+}
